### 2021-10-12 #3818 @whedon
[PRE REVIEW]: singcar: Comparing single cases to small samples in R
**Submitting author:** @jorittmo (<a href="http://orcid.org/0000-0001-5075-0166">Jonathan Ö. Rittmo</a>)
**Repository:** <a href="https://github.com/jorittmo/singcar" target ="_blank">https://github.com/jorittmo/singcar</a>
**Version:** v.0.1.3
