### 2021-10-20 #3818 @csoneson
👋🏻 @lindemann09, @Metafrancois - would you be interested in reviewing this submission for JOSS?

> singcar: Comparing single cases to small samples in R
