### 2021-11-25 jorittmo/singcar#2 @craddm
JOSS reviewer comments
Hi there,

This is my review for [JOSS](https://github.com/openjournals/joss-reviews/issues/3887).

It's a nice short one, I think. This seems to me a really nice well-documented package. Installed fine from both CRAN and Github, everything I've tried to do with it has worked, and it seems to have appropriate automated tests set up. The manuscript does a good job of covering the need for the package. All in all I think it's a really nice contribution.

The only thing I can comment on is regarding **Community Guidelines**, which are required by JOSS. Add some additional guidance for people who, as per the review criteria, wish to "1) Contribute to the software 2) Report issues or problems with the software 3) Seek support". See [here](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) - might be a bit overkill for your needs, but at least get a `contributing.md` with some simple info, and maybe add something in the `readme` file pointing people at the issue tracker etc. A good way to look for ideas is to look at some of the recently accepted JOSS articles through their issue tracker!

Cheers,
Matt
