### 2021-11-25 jorittmo/singcar#2 @jorittmo
Setting up some community guidelines is absolutely something I should do, just did not occur to me for some reason. Thank you very much for the review, I appreciate your work!

Best
