### 2021-12-14 jorittmo/singcar#2 @lindemann09
Hi,

This is my short review for [JOSS](https://github.com/openjournals/joss-reviews/issues/3887). Sorry for my delayed response.

With `singcar` your provide a very nice, well-documented and stable R packages. The manuscript is clear-cut.  I'm convinced that the package is highly needed and very useful for researcher and neuropsychologists working with single case data.

The only critical issue that I see has already been raised by @craddm (see above) and is related to the of missing community guidelines. I find it particularly relevant to clarify how to report issues or problems with the software (2).

cheers,
Oliver
