### 2021-12-16 386af4c
add community and contribution guidelines
