### 2021-12-16 jorittmo/singcar#2 @jorittmo
In commit [386af4c](https://github.com/jorittmo/singcar/commit/386af4c5913b4f7abd9a18cc98d79b2120a3c4bc) I have now addressed the lack of both contribution guidelines and a code of conduct by including a [contribution.md](https://github.com/jorittmo/singcar/blob/master/CONTRIBUTING.md) and [code_of_conduct.md](https://github.com/jorittmo/singcar/blob/master/CODE_OF_CONDUCT.md) (as well as pointing to them in the README).

Thank you both for the reviews!
