### 2021-12-28 bccc41e
update reference Mitchell et al. to include DOI
