+++ paper/paper.bib
@@ -132,7 +132,11 @@
 
 
 @article{mitchell2020peripheral,
-  title={Peripheral reaching in Alzheimer’s disease and mild cognitive impairment},
-  author={Mitchell, Alexandra G and Rossit, Stephanie and Pal, Suvankar and Hornberger, Michael and Warman, Annie and Kenning, Elise and Williamson, Laura and Shapland, Rebecca and McIntosh, Robert D},
-  year={2020}
+	Title = {Peripheral reaching in Alzheimer's disease and mild cognitive impairment},
+	Author = {Mitchell, Alexandra Grace and Rossit, Stephanie and Pal, Suvankar and Hornberger, Michael and Warman, Annie and Kenning, Elise and Williamson, Laura and Shapland, Rebecca and McIntosh, Robert},
+	DOI = {10.31234/osf.io/jbr2t},
+	Publisher = {PsyArXiv},
+	Year = {2020},
+	URL = {https://doi.org/10.31234/osf.io/jbr2t},
 }
+
