+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: singcar
 Title: Comparing Single Cases to Small Samples
-Version: 0.1.3.9000
+Version: 0.1.4
 Authors@R: 
     c(person(given = "Jonathan",
              family = "Rittmo",
