### 2021-12-29 #3887 @jorittmo
Hi @csoneson!

I've now made a tagged release ([v0.1.4](https://github.com/jorittmo/singcar/tree/v0.1.4)) and archived this version at Zenodo with DOI: [10.5281/zenodo.5808757](https://doi.org/10.5281/zenodo.5808757).

Best

> * [x]  Make a tagged release of the software, and list the version tag of the archived version here.
> * [x]  Archive the reviewed software in Zenodo or a similar service
> * [x]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
> * [x]  Please list the DOI of the archived version here.
